make formatting of ordered and unordered lists consistent
List were formatted inconsistently - items are ended with commas, dots or nothing. Items in ordered lists in most cases were ended with dots and in unordered with commas (the last one with a dot). This commit adds missing dot/comma character to all list items, so all lists are formatted according to the previously described rule (ordered - dot, unordered - comma)